+++ idlib/BitMsg.h
@@ -100,7 +100,7 @@ public:
 	int				ReadByte( void ) const;
 	int				ReadShort( void ) const;
 	int				ReadUShort( void ) const;
-	int				ReadLong( void ) const;
+	int				ReadInt( void ) const;
 	float			ReadFloat( void ) const;
 	float			ReadFloat( int exponentBits, int mantissaBits ) const;
 	float			ReadAngle8( void ) const;
@@ -363,7 +363,7 @@ ID_INLINE int idBitMsg::ReadUShort( void ) const {
 	return (unsigned short)ReadBits( 16 );
 }
 
-ID_INLINE int idBitMsg::ReadLong( void ) const {
+ID_INLINE int idBitMsg::ReadInt( void ) const {
 	return ReadBits( 32 );
 }
 
@@ -466,7 +466,7 @@ public:
 	int				ReadByte( void ) const;
 	int				ReadShort( void ) const;
 	int				ReadUShort( void ) const;
-	int				ReadLong( void ) const;
+	int				ReadInt( void ) const;
 	float			ReadFloat( void ) const;
 	float			ReadFloat( int exponentBits, int mantissaBits ) const;
 	float			ReadAngle8( void ) const;
@@ -609,7 +609,7 @@ ID_INLINE int idBitMsgDelta::ReadUShort( void ) const {
 	return (unsigned short)ReadBits( 16 );
 }
 
-ID_INLINE int idBitMsgDelta::ReadLong( void ) const {
+ID_INLINE int idBitMsgDelta::ReadInt( void ) const {
 	return ReadBits( 32 );
 }
